Rename StepIndicator as StepIndicatorElement named export
Make room for React implementation
